% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\docType{methods}
\name{dataRequest}
\alias{dataRequest}
\alias{dataRequest,character,character-method}
\title{Prepare a request to retrieve data from the drug register}
\usage{
dataRequest(source, idTable)

\S4method{dataRequest}{character,character}(source, idTable)
}
\arguments{
\item{source}{Data source, e.g. "LMR".}

\item{idTable}{A table ID.}
}
\value{
A dataframe.
}
\description{
Prepare a request to retrieve data from the drug register
}
\examples{
\dontrun{
library(getStatisticsFHI)
req <- dataRequest('lmr', '615')
req <- filterAge(req, filter='item', value='TOTALT')
getData(req)
}
}
