% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\docType{class}
\name{dataRequestAbstract-class}
\alias{dataRequestAbstract-class}
\title{An S4 class representing a data request}
\description{
Holds information regarding each request.
}
\section{Slots}{

\describe{
\item{\code{source}}{The data source.}

\item{\code{table}}{The table to retrieve data from.}

\item{\code{endpoint}}{The URL from which to retrieve data.}

\item{\code{filters}}{A list of filters to apply.}
}}

