% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\docType{methods}
\name{dataRequestLMR}
\alias{dataRequestLMR}
\title{Prepare a request to retrieve data from the drug register}
\usage{
dataRequestLMR(idTable)
}
\arguments{
\item{idTable}{A table ID.}
}
\value{
A dataframe.
}
\description{
Prepare a request to retrieve data from the drug register
}
\examples{
\dontrun{
library(getStatisticsFHI)
req <- dataRequestLMR('615')
req <- filterAge(req, filter='item', value='TOTALT')
getData(reg)
}
}
