% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{endpointTable}
\alias{endpointTable}
\title{Endpoint for a table.}
\usage{
endpointTable(source, idTable, type = "")
}
\arguments{
\item{source}{A source.}

\item{idTable}{A table ID.}

\item{type}{Type of information to retrieve.}
}
\description{
Endpoint for a table.
}
