% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{endpointTableData}
\alias{endpointTableData}
\title{Endpoint for table data.}
\usage{
endpointTableData(source, idTable)
}
\arguments{
\item{source}{A source.}

\item{idTable}{A table ID.}
}
\description{
Endpoint for table data.
}
