% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{endpointTableDesc}
\alias{endpointTableDesc}
\title{Endpoint for table description.}
\usage{
endpointTableDesc(source, idTable)
}
\arguments{
\item{source}{A source.}

\item{idTable}{A table ID.}
}
\description{
Endpoint for table description.
}
