% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{endpointTableMeta}
\alias{endpointTableMeta}
\title{Endpoint for meta information for a table.}
\usage{
endpointTableMeta(source, idTable)
}
\arguments{
\item{source}{A source.}

\item{idTable}{A table ID.}
}
\description{
Endpoint for meta information for a table.
}
