% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\docType{methods}
\name{filterAge}
\alias{filterAge}
\title{Filter data on age group.}
\usage{
filterAge(object, filter, value)
}
\arguments{
\item{object}{A dataRequest object.}

\item{filter}{Type of filter to apply: item, top or all.}

\item{value}{Filtering value, either an age or "TOTALT" for all groups.}
}
\value{
A dataRequest object with a filter applied.
}
\description{
Filter data on age group.
}
\examples{
\dontrun{
library(getStatisticsFHI)
req <- dataRequestLMR('615')
req <- filterAge(req, filter='item', value='TOTALT')
getData(req)
}
}
