% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\docType{methods}
\name{filterATC}
\alias{filterATC}
\title{Filter data on ATC code.}
\usage{
filterATC(object, filter, values)
}
\arguments{
\item{object}{A dataRequest object.}

\item{filter}{Type of filter to apply: item, top or all.}

\item{values}{Filtering value}
}
\value{
A dataRequest object with a filter applied.
}
\description{
Filter data on ATC code.
}
\examples{
\dontrun{
library(getStatisticsFHI)
req <- dataRequestLMR('615')
req <- filterATC(req, filter='all', value='AA*')
getData(req)
}
}
