% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getAgeGroup}
\alias{getAgeGroup}
\title{Function to simplify getting an age group by specifying an age you wish to look at.}
\usage{
getAgeGroup(age)
}
\arguments{
\item{age}{An age. If age is not an integer, age will be returned as is.}
}
\description{
Function to simplify getting an age group by specifying an age you wish to look at.
}
