% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\docType{methods}
\name{getData}
\alias{getData}
\alias{getData,dataRequestAbstract-method}
\title{Execute a request to get data.}
\usage{
getData(x, ...)

\S4method{getData}{dataRequestAbstract}(x, verb = 0)
}
\arguments{
\item{x}{A dataRequest object.}

\item{...}{Verbosity of execution.}

\item{verb}{Verbosity of execution.}
}
\value{
A dataframe.
}
\description{
Execute a request to get data.
}
\examples{
\dontrun{
library(getStatisticsFHI)
req <- dataRequestLMR('615')
req <- filterAge(req, filter='item', value='TOTALT')
getData(req)
}
}
