% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{getRequest}
\alias{getRequest}
\title{Perform a GET request.}
\usage{
getRequest(endpoint, dataFrame = T)
}
\arguments{
\item{endpoint}{An endpoint (returned by endpoint* functions).}

\item{dataFrame}{Whether to convert the response to a data.frame.}
}
\description{
Perform a GET request.
}
