% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{getSourceDesc}
\alias{getSourceDesc}
\title{Get description of a source.}
\usage{
getSourceDesc(source, dataFrame = T)
}
\arguments{
\item{source}{A source (e.g. lmr).}

\item{dataFrame}{Whether to convert the response to a data.frame.}
}
\description{
Get description of a source.
}
