% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{getSources}
\alias{getSources}
\title{Get available sources.}
\usage{
getSources(dataFrame = T)
}
\arguments{
\item{dataFrame}{Whether to convert the response to a data.frame.}
}
\description{
Get available sources.
}
