% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{getTableDesc}
\alias{getTableDesc}
\title{Get descriptoin of a table.}
\usage{
getTableDesc(source, idTable, dataFrame = T)
}
\arguments{
\item{source}{A source (e.g. lmr).}

\item{idTable}{The table ID.}

\item{dataFrame}{Whether to convert the response to a data.frame.}
}
\description{
Get descriptoin of a table.
}
