% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{getTableMeta}
\alias{getTableMeta}
\title{Get metadata for a table.}
\usage{
getTableMeta(source, idTable, dataFrame = T)
}
\arguments{
\item{source}{A source (e.g. lmr).}

\item{idTable}{The table ID.}

\item{dataFrame}{Whether to convert the response to a data.frame.}
}
\description{
Get metadata for a table.
}
