% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{recodeLMR}
\alias{recodeLMR}
\title{Function to recode drug register data into a better format.}
\usage{
recodeLMR(df)
}
\arguments{
\item{df}{A dataframe from a call to dataRequest('lmr', '615')}
}
\description{
Function to recode drug register data into a better format.
}
