% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\docType{methods}
\name{filters}
\alias{filters}
\title{Get filters that are applied to a request.}
\usage{
filters(object)
}
\arguments{
\item{object}{A dataRequest object.}
}
\value{
A list outlining the filters.
}
\description{
Get filters that are applied to a request.
}
\examples{
\dontrun{
library(getStatisticsFHI)
req <- dataRequestLMR('615')
req <- filterAge(req, filter='item', value='TOTALT')
filters(req)
}
}
