% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\docType{methods}
\name{getData}
\alias{getData}
\alias{getData,dataRequestLMR,logical,numeric-method}
\title{Execute a request to get data.}
\usage{
getData(x, recode, verb)

\S4method{getData}{dataRequestLMR,logical,numeric}(x, recode = F, verb = 0)
}
\arguments{
\item{x}{A dataRequest object.}

\item{recode}{Whether to transform raw data.}

\item{verb}{Verbosity of execution.}
}
\value{
A dataframe.
}
\description{
Execute a request to get data.
}
\examples{
\dontrun{
library(getStatisticsFHI)
req <- dataRequest('lmr', '615')
req <- filterAge(req, filter='item', value='TOTALT')
getData(req)
}
}
